Cleaner usage string

+++ source/tests/nuclear.cpp
@@ -85,7 +85,7 @@ int p3(int argc, char** argv)
     parser.add_flag('z', "param_z", "The parameter z");
     parser.add_variable<int>('m', "var_m", "The variable m", 10);
     parser.add_variable<int>('n', "var_n", "The variable n", 10);
-    parser.add_variable<int>('o', "var_o", "The variable o", 10);
+    parser.add_variable<float>('o', "var_o", "The variable o", 10);
     parser.add_positional<int>("MAGIC", "The magic number");
 
     parser.set_flags_exclusive({'x', 'y'});
